Asteroids, Shooting Mechanics, Movement & Physics Reworked, & More
-Initial asteroid implementation & spawning.

-Refactored movement system to use Unity built in physics system rather then manually simulating movement.

-Configured frame rate for better performance & added a frame rate controller.

-Added the groundwork for a UI, game  loop, & a Start & Game Over Screens.

-Fixed issues with physics system and added specific collision triggers.

-Added textures for a background.

-Added ship's ability to shoot plasma projectiles.

+++ Assets/Scripts/Health Bar/UpdateHealth.cs
@@ -15,7 +15,7 @@ public class UpdateHealth : MonoBehaviour
     {
         if (spaceShip != null)
         {   
-            var shipHealth = spaceShip.GetComponent<Health>().healthPoints;
+            var shipHealth = (int)spaceShip.GetComponent<Health>().healthPoints;
             if (shipHealth <= 0)
             {
                 spriteRenderer.sprite = null;
